requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
defusedxml==0.7.1
Django==3.2.5
django-countries==7.3.2
django-crispy-forms==1.14.0
django-dbbackup==4.0.1
django-filter==23.4
django-sms==0.5.0
django-widget-tweaks==1.5.0
docopt==0.6.2
fpdf2==2.5.6
idna==3.3
numpy==1.23.2
Pillow==9.2.0
pipreqs==0.4.11
PyJWT==2.4.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
stripe==4.0.2
svg.path==6.2
twilio==7.12.1
typing-extensions==4.3.0
urllib3==1.26.11
yarg==0.1.9
